chore: rename text to text_simple
