+++ crates/motiongfx_vello/src/vello_motion/mod.rs
@@ -1,4 +1,4 @@
 pub(crate) mod circle_motion;
 pub(crate) mod line_motion;
 pub(crate) mod rect_motion;
-pub(crate) mod text_motion;
+pub(crate) mod text_simple_motion;
